complete log and change Status to a copy-swap idom

+++ src/Util/Hash.cc
@@ -19,7 +19,7 @@ namespace CDB{
 	 *  @param
 	 *  @return a number indicate the hash Result;
 	 */
-	size_t Hash(const char *data,size_t n,uint32_t seed){
+	uint32_t Hash(const char *data,size_t n,uint32_t seed){
 		const uint32_t m = 0xc6a4a793;
 		const uint32_t r = 24;
 		const char* limit = data + n;
